requirements.txt
requests~=2.26.0
beautifulsoup4~=4.9.3
pyspellchecker==0.6.2